[HWKMETRICS-113] move classes into request package and add api docs
