+++ api/metrics-api-jaxrs/src/main/java/org/hawkular/metrics/api/jaxrs/request/MixedMetricsRequest.java
@@ -21,8 +21,6 @@ import java.util.List;
 
 import com.google.common.base.Objects;
 import com.wordnik.swagger.annotations.ApiModel;
-import org.hawkular.metrics.api.jaxrs.param.Availability;
-import org.hawkular.metrics.api.jaxrs.param.Gauge;
 
 /**
  * @author Stefan Negrea
