--- api/metrics-api-jaxrs/src/main/resources/results.json
@@ -1,18 +1,0 @@
-[
-  {
-    "id": "m14",
-    "tenantId": "T5858dab5-3db9-4fc3-977e-48adc794fdc910"
-  },
-  {
-    "id": "m15",
-    "tenantId": "T5858dab5-3db9-4fc3-977e-48adc794fdc910"
-  },
-  {
-    "id": "m16",
-    "tags": {
-      "a10": "10",
-      "a11": "11"
-    },
-    "tenantId": "T5858dab5-3db9-4fc3-977e-48adc794fdc910"
-  }
-]
